requirements.txt
psutil
Adafruit-SSD1306
Pillow
